make scripts were rewritten

+++ make_MagOS/files/patches/rootfs/MagOS/usr/lib/magos/rootfs-patches/MagOS-clean_distr.sh
@@ -30,14 +30,10 @@ rm -fr 2>/dev/null \
 
 
 #wallpapers
-for a in usr/share/wallpapers/* ;do
-    [ -d "$a" ] && rm -fr "$a"
-done
+[ -d usr/share/wallpapers ] && for a in usr/share/wallpapers/* ;do [ -d "$a" ] && rm -fr "$a" ; done
 
 #Kill big icons to save some space
-for a in `find usr/share/icons -type d | egrep -e '[x/]512$|[x/]256$|/128x128$' ` ;do
-   rm -fr "$a"
-done
+[ -d usr/share/icons ] && for a in `find usr/share/icons -type d | egrep -e '[x/]512$|[x/]256$|/128x128$' ` ;do  rm -fr "$a" ; done
 
 #KDE
 rm -fr 2>/dev/null usr/share/autostart/kaddressbookmigrator.desktop \
